requirements.txt
asgiref==3.6.0
black==23.1.0
click==8.1.3
Django==3.2.17
django-admin==2.0.2
django-excel-base==1.0.5
django-excel-response2==3.0.3
django-six==1.0.5
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycodestyle==2.10.0
pytz==2022.7.1
screen==1.0.1
sqlparse==0.4.3
tomli==2.0.1
xlwt==1.3.0
